@@ -17,7 +17,7 @@ class CarrierLoaded extends CarrierState {
 class CarriersLoaded extends CarrierState {
   const CarriersLoaded(this.carriers, this.carrierType);
   final List<Carrier> carriers;
-  final CarrierType carrierType;
+  final CarrierType? carrierType;
 }
 
 class CarrierFailure extends CarrierState {
