@@ -28,5 +28,5 @@ class DeleteOrder extends OrderEvent {
 
 class GetOrders extends OrderEvent {
   const GetOrders(this.orderStatus);
-  final OrderStatus orderStatus;
+  final OrderStatus? orderStatus;
 }
