@@ -28,5 +28,5 @@ class DeleteTenant extends TenantEvent {
 
 class GetTenants extends TenantEvent {
   const GetTenants(this.tenantType);
-  final TenantType tenantType;
+  final TenantType? tenantType;
 }
