@@ -33,5 +33,5 @@ class DeleteUser extends UserEvent {
 
 class GetUsers extends UserEvent {
   const GetUsers(this.userRole);
-  final UserRole userRole;
+  final UserRole? userRole;
 }
