@@ -13,9 +13,9 @@ import 'package:shipanther/extensions/order_extension.dart';
 class OrderAddEdit extends StatefulWidget {
   const OrderAddEdit(
     this.loggedInUser, {
-    @required this.order,
-    @required this.orderBloc,
-    @required this.isEdit,
+    required this.order,
+    required this.orderBloc,
+    required this.isEdit,
   });
   final api.User loggedInUser;
   final api.Order order;
@@ -29,10 +29,10 @@ class OrderAddEdit extends StatefulWidget {
 class _OrderAddEditState extends State<OrderAddEdit> {
   static final GlobalKey<FormState> formKey = GlobalKey<FormState>();
 
-  String _orderSerialNumber;
-  api.OrderStatus _orderStatus;
-  api.Tenant _tenant;
-  api.Customer _customer;
+  final TextEditingController _serialNumber = TextEditingController();
+  api.OrderStatus? _orderStatus;
+  api.Tenant? _tenant;
+  api.Customer? _customer;
   final TextEditingController _tenantTypeAheadController =
       TextEditingController();
   final TextEditingController _customerTypeAheadController =
@@ -50,10 +50,10 @@ class _OrderAddEditState extends State<OrderAddEdit> {
       appBar: AppBar(
         title: Text(
           widget.isEdit
-              ? ShipantherLocalizations.of(context)
-                  .editParam(ShipantherLocalizations.of(context).ordersTitle(1))
-              : ShipantherLocalizations.of(context).addNewParam(
-                  ShipantherLocalizations.of(context).ordersTitle(1)),
+              ? ShipantherLocalizations.of(context)!.editParam(
+                  ShipantherLocalizations.of(context)!.ordersTitle(1))
+              : ShipantherLocalizations.of(context)!.addNewParam(
+                  ShipantherLocalizations.of(context)!.ordersTitle(1)),
         ),
         centerTitle: true,
       ),
@@ -69,22 +69,22 @@ class _OrderAddEditState extends State<OrderAddEdit> {
           child: ListView(
             children: [
                   TextFormField(
-                    initialValue: widget.order.serialNumber ?? '',
+                    initialValue: widget.order.serialNumber,
                     autofocus: !widget.isEdit,
+                    controller: _serialNumber,
                     style: Theme.of(context).textTheme.headline5,
                     decoration: InputDecoration(
                         labelText:
-                            ShipantherLocalizations.of(context).orderNumber),
+                            ShipantherLocalizations.of(context)!.orderNumber),
                     maxLength: 15,
-                    validator: (val) => val.trim().isEmpty
-                        ? ShipantherLocalizations.of(context).paramEmpty(
-                            ShipantherLocalizations.of(context).orderNumber)
+                    validator: (val) => val == null || val.trim().isEmpty
+                        ? ShipantherLocalizations.of(context)!.paramEmpty(
+                            ShipantherLocalizations.of(context)!.orderNumber)
                         : null,
-                    onSaved: (value) => _orderSerialNumber = value,
                   ),
                   if (!widget.loggedInUser.isCustomer)
                     smartSelect<api.OrderStatus>(
-                      title: ShipantherLocalizations.of(context).orderStatus,
+                      title: ShipantherLocalizations.of(context)!.orderStatus,
                       onChange: (state) => _orderStatus = state.value,
                       choiceItems:
                           S2Choice.listFrom<api.OrderStatus, api.OrderStatus>(
@@ -120,24 +120,22 @@ class _OrderAddEditState extends State<OrderAddEdit> {
       ),
       floatingActionButton: FloatingActionButton(
         tooltip: widget.isEdit
-            ? ShipantherLocalizations.of(context).edit
-            : ShipantherLocalizations.of(context).create,
+            ? ShipantherLocalizations.of(context)!.edit
+            : ShipantherLocalizations.of(context)!.create,
         child: Icon(widget.isEdit ? Icons.check : Icons.add),
         onPressed: () async {
-          final form = formKey.currentState;
-          if (form.validate()) {
-            form.save();
-            widget.order.serialNumber = _orderSerialNumber;
+          if (formKey.currentState!.validate()) {
+            widget.order.serialNumber = _serialNumber.text;
             widget.order.status = _orderStatus ?? api.OrderStatus.open;
             if (widget.loggedInUser.isCustomer) {
               widget.order.customerId = widget.loggedInUser.customerId;
               _customer = null;
             }
             if (_customer != null) {
-              widget.order.customerId = _customer.id;
+              widget.order.customerId = _customer!.id;
             }
             if (_tenant != null) {
-              widget.order.tenantId = _tenant.id;
+              widget.order.tenantId = _tenant!.id;
             }
             widget.order.createdBy =
                 (await context.read<UserRepository>().self()).id;
@@ -155,6 +153,7 @@ class _OrderAddEditState extends State<OrderAddEdit> {
 
   @override
   void dispose() {
+    _serialNumber.dispose();
     _tenantTypeAheadController.dispose();
     _customerTypeAheadController.dispose();
     super.dispose();
