@@ -20,7 +20,7 @@ class ContainerScreen extends StatefulWidget {
 }
 
 class _ContainerScreenState extends State<ContainerScreen> {
-  ShipmentBloc bloc;
+  late ShipmentBloc bloc;
   @override
   void initState() {
     super.initState();
@@ -61,7 +61,7 @@ class _ContainerScreenState extends State<ContainerScreen> {
         return ShipantherScaffold(
           widget.loggedInUser,
           bottomNavigationBar: null,
-          title: ShipantherLocalizations.of(context).shipmentsTitle(2),
+          title: ShipantherLocalizations.of(context)!.shipmentsTitle(2),
           actions: const [],
           body: const CenteredLoading(),
           floatingActionButton: null,
