@@ -12,7 +12,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 class UserList extends StatelessWidget {
   const UserList(this.loggedInUser,
-      {Key key, @required this.userLoadedState, this.userBloc})
+      {Key? key, required this.userLoadedState, required this.userBloc})
       : super(key: key);
 
   final UserBloc userBloc;
@@ -21,7 +21,7 @@ class UserList extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final title = ShipantherLocalizations.of(context).usersTitle(2);
+    final title = ShipantherLocalizations.of(context)!.usersTitle(2);
     final actions = <Widget>[
       FilterButton<UserRole>(
         possibleValues: UserRole.values,
@@ -30,7 +30,7 @@ class UserList extends StatelessWidget {
         onSelected: (t) => context.read<UserBloc>().add(
               GetUsers(t),
             ),
-        tooltip: ShipantherLocalizations.of(context).userTypeFilter,
+        tooltip: ShipantherLocalizations.of(context)!.userTypeFilter,
       )
     ];
 
@@ -73,22 +73,22 @@ class UserList extends StatelessWidget {
               ),
               children: [
                 displaySubtitle(
-                    ShipantherLocalizations.of(context).email, t.email),
+                    ShipantherLocalizations.of(context)!.email, t.email),
                 displaySubtitle(
-                    ShipantherLocalizations.of(context).role, t.role.text),
+                    ShipantherLocalizations.of(context)!.role, t.role.text),
                 displaySubtitle(
-                    ShipantherLocalizations.of(context).createdAt,
-                    ShipantherLocalizations.of(context)
+                    ShipantherLocalizations.of(context)!.createdAt,
+                    ShipantherLocalizations.of(context)!
                         .dateFormatter
                         .format(t.createdAt)),
                 displaySubtitle(
-                    ShipantherLocalizations.of(context).lastUpdate,
-                    ShipantherLocalizations.of(context)
+                    ShipantherLocalizations.of(context)!.lastUpdate,
+                    ShipantherLocalizations.of(context)!
                         .dateFormatter
                         .format(t.updatedAt)),
                 if (loggedInUser.isSuperAdmin)
                   displaySubtitle(
-                      ShipantherLocalizations.of(context).tenantId, t.tenantId)
+                      ShipantherLocalizations.of(context)!.tenantId, t.tenantId)
                 else
                   const Text(''),
               ],
