@@ -5,6 +5,6 @@ abstract class CarrierRepository {
   Future<Carrier> fetchCarrier(String id);
   Future<Carrier> createCarrier(Carrier carrier);
   Future<Carrier> updateCarrier(String id, Carrier carrier);
-  Future<List<Carrier>> fetchCarriers();
-  Future<List<Carrier>> filterCarriers(CarrierType carrierType);
+  Future<List<Carrier>> fetchCarriers(
+      {int? page = 1, CarrierType? carrierType, String? name});
 }
