@@ -5,6 +5,9 @@ abstract class OrderRepository {
   Future<Order> fetchOrder(String id);
   Future<Order> createOrder(Order order);
   Future<Order> updateOrder(String id, Order order);
-  Future<List<Order>> fetchOrders();
-  Future<List<Order>> filterOrders(OrderStatus orderStatus);
+  Future<List<Order>> fetchOrders(
+      {int? page = 1,
+      OrderStatus? orderStatus,
+      String? customerId,
+      String? serialNumber});
 }
