@@ -25,14 +25,31 @@ class RemoteShipmentRepository extends ShipmentRepository {
   }
 
   @override
-  Future<List<Shipment>> fetchShipments() async {
+  Future<List<Shipment>> fetchShipments(
+      {int? page = 1,
+      ShipmentType? shipmentType,
+      ShipmentStatus? shipmentStatus,
+      ShipmentSize? shipmentSize,
+      String? terminalId,
+      String? carrierId,
+      String? driverId,
+      String? orderId,
+      String? serialNumber}) async {
     final client = await _apiRepository.apiClient();
-    return client.shipmentsGet();
+    return client.shipmentsGet(
+        page: page,
+        type: shipmentType,
+        size: shipmentSize,
+        status: shipmentStatus,
+        //terminalId: terminalId,
+        driverId: driverId,
+        orderId: orderId,
+        serialNumber: serialNumber,
+        carrierId: carrierId);
   }
 
   @override
-  Future<List<Shipment>> filterShipments(
-      ShipmentStatus shipmentStatus) async {
+  Future<List<Shipment>> filterShipments(ShipmentStatus shipmentStatus) async {
     final shipments = await fetchShipments();
     if (shipmentStatus == null) {
       return shipments;
