@@ -13,9 +13,11 @@ class RemoteTerminalRepository extends TerminalRepository {
   }
 
   @override
-  Future<List<Terminal>> fetchTerminals() async {
+  Future<List<Terminal>> fetchTerminals(
+      {int? page = 1, TerminalType? terminalType, String? name}) async {
     final client = await _apiRepository.apiClient();
-    return await client.terminalsGet();
+    return await client.terminalsGet(
+        page: page, type: terminalType, name: name);
   }
 
   @override
