@@ -5,6 +5,6 @@ abstract class TerminalRepository {
   Future<Terminal> fetchTerminal(String id);
   Future<Terminal> createTerminal(Terminal terminal);
   Future<Terminal> updateTerminal(String id, Terminal terminal);
-  Future<List<Terminal>> fetchTerminals();
-  Future<List<Terminal>> filterTerminals(TerminalType terminalType);
+  Future<List<Terminal>> fetchTerminals(
+      {int? page = 1, TerminalType? terminalType, String? name});
 }
